+++ java-client/src/main/java/org/opensearch/client/transport/client_metrics/MetricConstants.java
@@ -1,0 +1,21 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.client.transport.client_metrics;
+
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.Metrics;
+import java.util.EnumSet;
+import java.util.Set;
+
+public class MetricConstants {
+    public static final double[] DEFAULT_PERCENTILES = new double[] { 0.99, 0.95, 0.9, 0.75, 0.5 };
+    public static final MeterRegistry DEFAULT_REGISTRY = Metrics.globalRegistry;
+    public static final Set<MetricTag> DEFAULT_EXCLUDED_TAGS = EnumSet.noneOf(MetricTag.class);
+    public static final Set<MetricGroup> DEFAULT_ADDITIONAL_METRIC_GROUPS = EnumSet.noneOf(MetricGroup.class);
+}